+++ experimental/packages/otlp-exporter-base/src/configuration/convert-legacy-node-http-options.ts
@@ -71,6 +71,7 @@ export function convertLegacyHttpOptions(
       compression: config.compression,
       agentFactory: convertLegacyAgentOptions(config),
       userAgent: config.userAgent,
+      fetch: config.fetch,
     },
     getNodeHttpConfigurationFromEnvironment(
       signalIdentifier,